+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased
 
+* Support `RUSTFLAGS` "deny warnings" in cargo zng.
+* Warn when `.zr-copy` does not find the directory or file.
+* Refactor `.zr-apk` to not require to be inside the staging dir.
+* Refactor `Impl Future` parameters into `impl IntoFuture`. 
+* Implement `IntoFuture for ResponseVar<T>`.
+* Remove `.zr-apk` requirement of extension on the folder name.
 
 # 0.11.1